+++ xmppserver/src/main/java/org/jivesoftware/openfire/sasl/SaslServerFactoryImpl.java
@@ -53,8 +53,8 @@ public class SaslServerFactoryImpl implements SaslServerFactory
         allMechanisms = new HashSet<>();
         allMechanisms.add( new Mechanism( "ANONYMOUS", true, true ) );
         allMechanisms.add( new Mechanism( "PLAIN", false, true ) );
-        allMechanisms.add( new Mechanism( "SCRAM-SHA-1", false, false ) );
-        allMechanisms.add( new Mechanism( "SCRAM-SHA-1-PLUS", false, false ) );
+        allMechanisms.add( new Mechanism( ScramSha1SaslServer.MECHANISM_NAME, false, false ) );
+        allMechanisms.add( new Mechanism( ScramSha1SaslServer.MECHANISM_NAME + "-PLUS", false, false ) );
         allMechanisms.add( new Mechanism( "JIVE-SHAREDSECRET", true, false ) );
         allMechanisms.add( new Mechanism( "EXTERNAL", false, false ) );
     }
@@ -78,10 +78,10 @@ public class SaslServerFactoryImpl implements SaslServerFactory
                 }
                 return new SaslServerPlainImpl( protocol, serverName, props, cbh );
 
-            case "SCRAM-SHA-1":
+            case ScramSha1SaslServer.MECHANISM_NAME:
                 return new ScramSha1SaslServer(false, props);
 
-            case "SCRAM-SHA-1-PLUS":
+            case ScramSha1SaslServer.MECHANISM_NAME + "-PLUS":
                 return new ScramSha1SaslServer(true, props);
 
             case "ANONYMOUS":
